Precision is important

+++ Day08/day08.cpp
@@ -200,7 +200,7 @@ int main()
     Helper helper;
     helper.calculateAnswers(inputFileName);
 
-    int answer = helper.getFirstPuzzleAnswer();
+    unsigned long long answer = helper.getFirstPuzzleAnswer();
     std::cout << "First half answer: " << answer << std::endl;
 
     answer = helper.getSecondPuzzleAnswer();
